+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Make compatible with a patched markdown-it-emoji ([#294](https://github.com/marp-team/marp-core/pull/294))
+
 ## v2.4.1 - 2022-04-12
 
 ### Changed